Have GH tests running for strict and devmode (#2545)
* Have gh tests running for strict and devmode

* Have gh tests running on pull and push regardless the branch

* More fixes on GH tests

* Adding home-read-all interface

* Remove CA_CERT references

* Remove run_with_sudo references

* Allow dbctl and helm to see your home

* Adding some debug messages

* Create an inspection tarball in case of failure

* Style fixes

* Enable and disable to read your home

* Fix dbctl command denial

* Try fixing helm and install hook ifconfig denials

* Connect personal files in test

* Connect kernel module control interface

* Helm unde dot-kube

* Connect interfaces in devmode

* Fix the smoketest

* Remove debugging output and upgrade tests

* Clean helm and kube directories

* Why do we exit 1 in a bash script?

* Gather inspect reports from GH actions

* Name the inspection reports placeholders correctly

* Fix the test-cluster test

* Fix style in cluster testing

* Fixes in the inspect command

* Adding two more interfaces

* Move list of interfaces to connect under utils.py

* Require connecting log-observe and login-session-observe

* Make some more interfaces required

* Tidying up

+++ scripts/inspect.sh
@@ -16,9 +16,9 @@ function check_service {
   # Check the service passed as the first argument is up and running and collect its logs.
   local service=$1
   mkdir -p $INSPECT_DUMP/$service
-  journalctl -n $JOURNALCTL_LIMIT -u $service &> $INSPECT_DUMP/$service/journal.log
-  systemctl status $service &> $INSPECT_DUMP/$service/systemctl.log
-  if systemctl status $service &> /dev/null
+  journalctl -n $JOURNALCTL_LIMIT -u snap.$service &> $INSPECT_DUMP/$service/journal.log
+  snapctl services $service &> $INSPECT_DUMP/$service/snapctl.log
+  if snapctl services $service | grep active &> /dev/null
   then
     printf -- '  Service %s is running\n' "$service"
   else
@@ -32,12 +32,7 @@ function check_service {
 function check_apparmor {
   # Collect apparmor info.
   mkdir -p $INSPECT_DUMP/apparmor
-  if [ -f /etc/apparmor.d/containerd ]
-  then
-    cp /etc/apparmor.d/containerd $INSPECT_DUMP/apparmor/
-  fi
   dmesg &> $INSPECT_DUMP/apparmor/dmesg
-  aa-status &> $INSPECT_DUMP/apparmor/aa-status
 }
 
 
@@ -66,12 +61,6 @@ function store_sys {
   # collect the processes running
   printf -- '  Copy processes list to the final report tarball\n'
   ps -ef > $INSPECT_DUMP/sys/ps
-  printf -- '  Copy snap list to the final report tarball\n'
-  snap version > $INSPECT_DUMP/sys/snap-version
-  snap list > $INSPECT_DUMP/sys/snap-list
-  # Stores VM name (or none, if we are not on a VM)
-  printf -- '  Copy VM name (or none) to the final report tarball\n'
-  systemd-detect-virt &> $INSPECT_DUMP/sys/vm_name
   # Store disk usage information
   printf -- '  Copy disk usage information to the final report tarball\n'
   df -h | grep ^/ &> $INSPECT_DUMP/sys/disk_usage # remove the grep to also include virtual in-memory filesystems
@@ -81,9 +70,6 @@ function store_sys {
   # Store server's uptime.
   printf -- '  Copy server uptime to the final report tarball\n'
   uptime &> $INSPECT_DUMP/sys/uptime
-  # Store the current linux distro.
-  printf -- '  Copy current linux distribution to the final report tarball\n'
-  lsb_release -a &> $INSPECT_DUMP/sys/lsb_release
   # Store openssl information.
   printf -- '  Copy openSSL information to the final report tarball\n'
   openssl version -v -d -e &> $INSPECT_DUMP/sys/openssl
@@ -94,12 +80,12 @@ function store_kubernetes_info {
   # Collect some in-k8s details
   printf -- '  Inspect kubernetes cluster\n'
   mkdir -p $INSPECT_DUMP/k8s
-  sudo -E /snap/bin/microk8s kubectl version 2>&1 | sudo tee $INSPECT_DUMP/k8s/version > /dev/null
-  sudo -E /snap/bin/microk8s kubectl cluster-info 2>&1 | sudo tee $INSPECT_DUMP/k8s/cluster-info > /dev/null
-  sudo -E /snap/bin/microk8s kubectl cluster-info dump -A 2>&1 | sudo tee $INSPECT_DUMP/k8s/cluster-info-dump > /dev/null
-  sudo -E /snap/bin/microk8s kubectl get all --all-namespaces -o wide 2>&1 | sudo tee $INSPECT_DUMP/k8s/get-all > /dev/null
-  sudo -E /snap/bin/microk8s kubectl get pv 2>&1 | sudo tee $INSPECT_DUMP/k8s/get-pv > /dev/null # 2>&1 redirects stderr and stdout to /dev/null if no resources found
-  sudo -E /snap/bin/microk8s kubectl get pvc 2>&1 | sudo tee $INSPECT_DUMP/k8s/get-pvc > /dev/null # 2>&1 redirects stderr and stdout to /dev/null if no resources found
+  /snap/bin/microk8s kubectl version 2>&1 | tee $INSPECT_DUMP/k8s/version > /dev/null
+  /snap/bin/microk8s kubectl cluster-info 2>&1 | tee $INSPECT_DUMP/k8s/cluster-info > /dev/null
+  /snap/bin/microk8s kubectl cluster-info dump -A 2>&1 | tee $INSPECT_DUMP/k8s/cluster-info-dump > /dev/null
+  /snap/bin/microk8s kubectl get all --all-namespaces -o wide 2>&1 | tee $INSPECT_DUMP/k8s/get-all > /dev/null
+  /snap/bin/microk8s kubectl get pv 2>&1 | tee $INSPECT_DUMP/k8s/get-pv > /dev/null # 2>&1 redirects stderr and stdout to /dev/null if no resources found
+  /snap/bin/microk8s kubectl get pvc 2>&1 | tee $INSPECT_DUMP/k8s/get-pvc > /dev/null # 2>&1 redirects stderr and stdout to /dev/null if no resources found
 }
 
 
@@ -107,9 +93,9 @@ function store_juju_info {
   # Collect some juju details
   printf -- '  Inspect Juju\n'
   mkdir -p $INSPECT_DUMP/juju
-  sudo -E /snap/bin/microk8s juju status 2>&1 | sudo tee $INSPECT_DUMP/juju/status > /dev/null
-  sudo -E /snap/bin/microk8s juju debug-log 2>&1 | sudo tee $INSPECT_DUMP/juju/debug.log > /dev/null
-  sudo -E /snap/bin/microk8s kubectl logs -n controller-uk8s --tail 10000 -c api-server controller-0 2>&1 | sudo tee $INSPECT_DUMP/juju/controller.log > /dev/null
+  /snap/bin/microk8s juju status 2>&1 | tee $INSPECT_DUMP/juju/status > /dev/null
+  /snap/bin/microk8s juju debug-log 2>&1 | tee $INSPECT_DUMP/juju/debug.log > /dev/null
+  /snap/bin/microk8s kubectl logs -n controller-uk8s --tail 10000 -c api-server controller-0 2>&1 | tee $INSPECT_DUMP/juju/controller.log > /dev/null
 }
 
 
@@ -117,15 +103,15 @@ function store_kubeflow_info {
   # Collect some kubeflow details
   printf -- '  Inspect Kubeflow\n'
   mkdir -p $INSPECT_DUMP/kubeflow
-  sudo -E /snap/bin/microk8s kubectl get pods -nkubeflow -oyaml 2>&1 | sudo tee $INSPECT_DUMP/kubeflow/pods.yaml > /dev/null
-  sudo -E /snap/bin/microk8s kubectl describe pods -nkubeflow 2>&1 | sudo tee $INSPECT_DUMP/kubeflow/pods.describe > /dev/null
+  /snap/bin/microk8s kubectl get pods -nkubeflow -oyaml 2>&1 | tee $INSPECT_DUMP/kubeflow/pods.yaml > /dev/null
+  /snap/bin/microk8s kubectl describe pods -nkubeflow 2>&1 | tee $INSPECT_DUMP/kubeflow/pods.describe > /dev/null
 }
 
 
 function suggest_fixes {
   # Propose fixes
   printf '\n'
-  if ! systemctl status snap.microk8s.daemon-kubelite &> /dev/null
+  if ! snapctl services $service | grep active &> /dev/null
   then
     if lsof -Pi :16443 -sTCP:LISTEN -t &> /dev/null
     then
@@ -230,19 +216,6 @@ function suggest_fixes {
     fi
   fi
 
-  # LXD Specific Checks
-  if cat /proc/1/environ | grep "container=lxc" &> /dev/null
-    then
-
-    # make sure the /dev/kmsg is available, indicating a potential missing profile
-    if [ ! -c "/dev/kmsg" ]  # kmsg is a character device
-    then
-      printf -- '\033[0;33mWARNING: \033[0m the lxc profile for MicroK8s might be missing. \n'
-      printf -- '\t  Refer to this help document to get MicroK8s working in with LXD: \n'
-      printf -- '\t  https://microk8s.io/docs/lxd \n'
-    fi
-  fi
-
   # node name
   nodename="$(hostname)"
   if [[ "$nodename" =~ [A-Z|_] ]] && ! grep -e "hostname-override" /var/snap/microk8s/current/args/kubelet &> /dev/null
@@ -292,6 +265,11 @@ if [ ${#@} -ne 0 ] && [ "$*" == "--help" ]; then
   exit 0;
 fi;
 
+if [ "$EUID" -ne 0 ]
+  then echo "Please run the inspection script with sudo"
+  exit 1
+fi
+
 rm -rf ${SNAP_DATA}/inspection-report
 mkdir -p ${SNAP_DATA}/inspection-report
 
@@ -299,24 +277,24 @@ printf -- 'Inspecting Certificates\n'
 check_certificates
 
 printf -- 'Inspecting services\n'
-check_service "snap.microk8s.daemon-cluster-agent"
-check_service "snap.microk8s.daemon-containerd"
-check_service "snap.microk8s.daemon-apiserver-kicker"
+check_service "microk8s.daemon-cluster-agent"
+check_service "microk8s.daemon-containerd"
+check_service "microk8s.daemon-apiserver-kicker"
 if [ -e "${SNAP_DATA}/var/lock/lite.lock" ]
 then
-  check_service "snap.microk8s.daemon-kubelite"
+  check_service "microk8s.daemon-kubelite"
 else
-  check_service "snap.microk8s.daemon-apiserver"
-  check_service "snap.microk8s.daemon-proxy"
-  check_service "snap.microk8s.daemon-kubelet"
-  check_service "snap.microk8s.daemon-scheduler"
-  check_service "snap.microk8s.daemon-controller-manager"
-  check_service "snap.microk8s.daemon-control-plane-kicker"
+  check_service "microk8s.daemon-apiserver"
+  check_service "microk8s.daemon-proxy"
+  check_service "microk8s.daemon-kubelet"
+  check_service "microk8s.daemon-scheduler"
+  check_service "microk8s.daemon-controller-manager"
+  check_service "microk8s.daemon-control-plane-kicker"
 fi
 if ! [ -e "${SNAP_DATA}/var/lock/ha-cluster" ]
 then
-  check_service "snap.microk8s.daemon-flanneld"
-  check_service "snap.microk8s.daemon-etcd"
+  check_service "microk8s.daemon-flanneld"
+  check_service "microk8s.daemon-etcd"
 fi
 
 store_args
@@ -331,15 +309,17 @@ store_network
 printf -- 'Inspecting kubernetes cluster\n'
 store_kubernetes_info
 
-printf -- 'Inspecting juju\n'
-store_juju_info
-
-printf -- 'Inspecting kubeflow\n'
-store_kubeflow_info
-
+### Juju and kubeflow are out of scope of the strict confinement work for now
+#
+#printf -- 'Inspecting juju\n'
+#store_juju_info
+#
+#printf -- 'Inspecting kubeflow\n'
+#store_kubeflow_info
+#
 suggest_fixes
-
-printf -- 'Building the report tarball\n'
+#
+#printf -- 'Building the report tarball\n'
 build_report_tarball
 
 exit $RETURN_CODE
